backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.116.1
fastapi_cors==0.0.6
firebase_admin==7.0.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
iniconfig==2.1.0
Jinja2==3.1.6
limits==5.4.0
loguru==0.7.3
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.47.1
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
wrapt==1.17.2
